Fixed typos and duplicate files, now executes
