+++ client/src/components/ChannelContainer.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+const ChannelContainer =() => {
+    return (
+        <div>
+            ChannelContainer
+
+        </div>
+    );
+}
+ export default ChannelContainer;
