+++ client/src/components/ChannelListContainer.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const ChannelListContainer =() => {
+    return (
+        <div>ChannelListContainer
+
+        </div>
+    );
+}
+ export default ChannelListContainer;
