+++ client/src/components/index.js
@@ -1,0 +1,2 @@
+export {default as ChannelContainer} from './ChannelContainer';
+export {default as ChannelListContainer} from './ChannelListContainer';
