+++ client/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+
 import App from './App';
-import './index.css';
+
 
 ReactDOM.render(<App/>, document.getElementById('root'));
