--- components/ChannelContainer.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-const ChannelContainer =() => {
-    return (
-        <div>ChannelContainer
-
-        </div>
-    );
-}
- export default ChannelContainer;
