--- components/ChannellistContainer.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-const ChannellistContainer =() => {
-    return (
-        <div>ChannellistContainer
-
-        </div>
-    );
-}
- export default ChannellistContainer;
