--- components/index.js
@@ -1,4 +1,0 @@
-
-//All the imports are being  done from here on out in one line 
-export{default as ChannelContainer} from './ChannelContainer';
-export {ChannelContainer} from './ChannellistContainer';
